add missing metadata titles and turn client pages into components
